Remove attachment app code
