+++ src/grm/urls.py
@@ -31,7 +31,6 @@ handler500 = "dashboard.authentication.views.handler500"
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("attachments/", include("attachments.urls")),
     path("authentication/", include("authentication.urls")),
     path("issues/", include("issues.urls")),
     path("i18n/", include("django.conf.urls.i18n")),
